+++ README.md
@@ -4,6 +4,8 @@ This repo document my practice in leet code problems.
 Problems are well catagorized (and easy for searching and learning).
 Some great solutions and some notes are also documented for each question.
 
+
+
 ## Instructions
 
 ### 1. Run execute_solution.py
@@ -11,13 +13,21 @@ In the execute_solution.py, you can see the dictionary of problems.
 The key is the name of the problem, and the value is the corresponding solution function of the problem.
 The user can change the problem name (based on the name in dictionary) in funtion main() to execute the selected function.
 
-
 ### 2. Run quick_run.py under folder quick_test
 During reviewing the questions, the user might want to re-practice the problem.
 The user can simply write the answer in the test_run function (and of couse change the input variable) and then execute this file.
 
 
+
 ## Good Learning Source while My Practice
+
 1. understand how differnt BigO looks like: https://stackoverflow.com/questions/2307283/what-does-olog-n-mean-exactly
 2. Neetcode: https://neetcode.io/practice
 3. LeedCode Top Interview: https://leetcode.com/problem-list/top-interview-questions/
+
+
+
+## Miscellaneous Note
+
+1. When deal with array problem, O(n^2) solution might be able to improve via Hashing or Sorting.
+2. Usually require pointers and while loop when facing Two Pointer or Sliding Window problems
