+++ execute_solution.py
@@ -1,4 +1,7 @@
-from problems.Array_and_Hashing.essential import ez__two_sum, ez__contain_duplicate
+from problems.array_and_hashing.essential import ez__two_sum, ez__contain_duplicate, ez__valid_anagram
+from problems.two_pointers.essential import ez_valid_palindrome
+from problems.sliding_window.essential import ez_best_time_to_buy_and_sell_stock
+
 from solutions import *
 
 from timeit import default_timer as timer
@@ -24,6 +27,14 @@ def solution_input():
                  # === Array and Hashing ===
                  'two_sum': [ez__two_sum.TwoSum.great_ans1, [2, 7, 11, 15], 18],    # imp: ***
                  'contains_duplicate': [ez__contain_duplicate.ContainsDuplicate.great_ans1, [0,0,1,2,3]],    # imp: **   
+                 'valid_anagram': [ez__valid_anagram.ValidAnagram.great_ans1, "anagram", "nagaram"],    # imp: ***
+
+                 # === Two Pointers ===
+                 'vali_palindrome': [ez_valid_palindrome.ValidPalindrome.great_ans1, "A man, a plan, a canal: Panama"],    # imp: ***
+
+                 # === Sliding Window ===
+                 'best_time_to_buy_and_sell_stock': [ez_best_time_to_buy_and_sell_stock.BestTimeToBuyAndSellStock.great_ans1, [7,1,5,3,6,4]],
+                 
 
                  'palindrome_number_0': [PalindromeNumber.my_solution, 121],
                  'palindrome_number_1': [PalindromeNumber.great_ans1, 122],
@@ -42,19 +53,8 @@ def solution_input():
 
                  'merge_two_sorted_lists_0': [MergeTwoSortedLists.my_solution, transfer_list_to_linked_nodes([1,2,4]), transfer_list_to_linked_nodes([1,3,4])],
 
-                 'vali_palindrome_0': [ValidPalindrome.my_solution, "A man, a plan, a canal: Panama"],
-                 'vali_palindrome_1': [ValidPalindrome.great_ans1, "A man, a plan, a canal: Panama"],
-                 'vali_palindrome_2': [ValidPalindrome.great_ans2, "A man, a plan, a canal: Panama"],
-
-                 'valid_anagram_0': [ValidAnagram.my_solution, "rat", "car"],
-                 'valid_anagram_1': [ValidAnagram.great_ans1, "anagram", "nagaram"],
-                 'valid_anagram_2': [ValidAnagram.great_ans2, "anagram", "nagaram"],
-
                  'binary_search_0': [BinarySearch.my_solution, [-1,0,3,5,9,12], 2],
                  'binary_search_1': [BinarySearch.great_ans1, [-1,0,3,5,9,12], 2],
-
-                 'best_time_to_buy_and_sell_stock_0': [BestTimeToBuyAndSellStock.my_solution, [7,1,5,3,6,4]],
-                 'best_time_to_buy_and_sell_stock_1': [BestTimeToBuyAndSellStock.great_ans1, [7,1,5,3,6,4]],
 
                  'single_number_0': [SingleNumber.my_solution, [2,2,1]],
                  'single_number_1': [SingleNumber.great_ans1, [2,2,1]],
@@ -71,5 +71,5 @@ def main(q_name):
 
 if __name__ == '__main__':
     # input question name you want to execute as argument
-    main('contains_duplicate')
+    main('best_time_to_buy_and_sell_stock')
