+++ solutions.py
@@ -259,150 +259,6 @@ class MergeTwoSortedLists:
 
         return transfer_linked_nodes_to_list(dummy.next) # if run on leetcode server, then simply return dummy.next
 
-
-        
-
-class ValidPalindrome:
-    """
-    A phrase is a palindrome if, after converting all uppercase letters into lowercase letters and 
-    removing all non-alphanumeric characters, it reads the same forward and backward. 
-    Alphanumeric characters include letters and numbers.
-
-    Given a string s, return true if it is a palindrome, or false otherwise.
-
-
-    Input: s = "A man, a plan, a canal: Panama"
-    Output: true
-    Explanation: "amanaplanacanalpanama" is a palindrome.
-    """
-
-    @staticmethod
-    def my_solution(s: str) -> bool:
-        """
-        tech use : two pointer
-        """
-        alpha_numeric = [char for char in "abcdefghijklmnopqrstuvwxyz0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ"]
-        s = s.lower()
-        lst_s = [char for char in s if char in alpha_numeric]
-
-        i = 0
-        while len(lst_s) > 1 and i < len(lst_s)-1:
-            if lst_s[i] != lst_s.pop():
-                return False
-            i += 1
-
-        return True
-
-
-
-    @staticmethod
-    def great_ans1(s: str) -> bool:
-        """
-        tech use : two pointer
-        https://www.youtube.com/watch?v=jJXJ16kPFWg
-        """
-
-        def _alphaNum(char):
-            # ord check the ASCII index of the character
-            return (ord('A') <= ord(char) <= ord('Z') or 
-                    ord('a') <= ord(char) <= ord('z') or 
-                    ord('0') <= ord(char) <= ord('9'))
-
-        lst_s = [char.lower() for char in s if _alphaNum(char)]
-        l, r = 0, len(lst_s)-1
-        while l < r:
-            if lst_s[l] != lst_s[r]:
-                return False
-            l += 1
-            r -= 1
-
-        # if don't tackle alphanumeric in beginning:
-        # while l < r:
-        #     while l < r and not _alphaNum(s[l]): l += 1
-        #     while l < r and not _alphaNum(s[r]): r -= 1
-        #     if s[l].lower() != s[r].lower():
-        #         return False
-        
-        return True
-    
-
-    @staticmethod
-    def great_ans2(s: str) -> bool:
-        lst_s = [char.lower() for char in s if char.isalnum()] 
-        return lst_s == lst_s[::-1]
-        
-
-class ValidAnagram:
-    """
-    Given two strings s and t, return true if t is an anagram of s, and false otherwise.
-    An Anagram is a word or phrase formed by rearranging the letters of a different word or phrase, 
-    typically using all the original letters exactly once.
-
-    Input: s = "anagram", t = "nagaram"
-    Output: true
-    """
-
-    @staticmethod
-    def my_solution(s: str, t: str) -> bool:
-        """
-        tech use: sort / hash table
-        """
-        # using sort:
-        # s_sort = [i for i in s]
-        # t_sort = [i for i in t]
-        # s_sort.sort() 
-        # t_sort.sort()
-        # return s_sort == t_sort
-
-        # without sort:
-        def _create_hash(str_input):
-            str_hash = {}
-            for i in str_input:
-                if i not in str_hash:
-                    str_hash[i] = 1
-                else:
-                    str_hash[i] += 1
-            return str_hash
-
-        s_hash = _create_hash(s)
-        t_hash = _create_hash(t)
-
-        return s_hash == t_hash
-
-
-    @staticmethod
-    def great_ans1(s: str, t: str) -> bool:
-        """
-        tech use : hash table
-        https://www.youtube.com/watch?v=9UtInBqnCgA
-        """
-        # following codes can be replaced by this in python:
-        # Counter(s) == Counter(t)
-
-        if len(s) != len(t):
-            return False
-        
-        hash_s, hash_t = {}, {}
-
-        for i in range(len(s)):
-            hash_s[s[i]] = hash_s.get(s[i], 0) + 1   # .get(s[i], 0) --> return 0 if the key doesn't exist yet
-            hash_t[t[i]] = hash_t.get(t[i], 0) + 1
-        
-        # here can be replaced by:  return hash_s == hash_t
-        for key in hash_s:
-            if hash_s[key] != hash_t.get(key, 0):
-                return False
-
-        return True
-    
-
-    @staticmethod
-    def great_ans2(s: str, t: str) -> bool:
-        """
-        tech use : sort
-        https://www.youtube.com/watch?v=9UtInBqnCgA
-        """
-        return sorted(s) == sorted(t)
         
 
 class BinarySearch:
@@ -457,71 +313,6 @@ class BinarySearch:
             else:
                 return mid
         return -1    
-
-
-class BestTimeToBuyAndSellStock:
-    """
-    You are given an array prices where prices[i] is the price of a given stock on the ith day.
-    You want to maximize your profit by choosing a single day to buy one stock and choosing a different day in the future to sell that stock.
-    Return the maximum profit you can achieve from this transaction. If you cannot achieve any profit, return 0.
-
-    Input: prices = [7,1,5,3,6,4]
-    Output: 5
-    Explanation: Buy on day 2 (price = 1) and sell on day 5 (price = 6), profit = 6-1 = 5.
-    Note that buying on day 2 and selling on day 1 is not allowed because you must buy before you sell.
-    """
-
-    @staticmethod
-    def my_solution(prices: List[int]) -> int:
-        
-        # --- Time out ---
-        # max_profit = 0 
-        # for i in range(len(prices)-1):
-        #     profit = max(prices[i:]) - prices[i]
-        #     if profit > max_profit:
-        #         max_profit = profit
-        # return max_profit
-
-        # --- Also Time out ---
-        if len(prices) <= 1:
-            return 0
-
-        l, r = 0, 1
-        max_p = 0
-
-        while r < len(prices):
-            if prices[l] < prices[r]:
-                r2 = r
-                while r2 < len(prices):
-                    profit = prices[r2] - prices[l]
-                    if profit > max_p:
-                        max_p = profit
-                    r2 += 1
-
-            l, r = l+1, r+1 
-
-        return max_p   
-
-            
-
-    @staticmethod
-    def great_ans1(prices: List[int]) -> int:
-        """
-        tech use : sliding window
-        https://www.youtube.com/watch?v=1pkOgXD63yU
-        """
-        l, r = 0, 1
-        max_p = 0
-
-        while r < len(prices):
-            if prices[l] < prices[r]:
-                max_p = max(prices[r] - prices[l], max_p)
-            else:
-                l = r   # shift to new lowest point 
-            
-            r += 1
-
-        return max_p   
 
 
 class SingleNumber:
